+++ authority-portal-frontend/src/app/core/api/fake-backend/impl/fake-users.ts
@@ -2,6 +2,7 @@ import {UserInfo} from '@sovity.de/authority-portal-client';
 
 export let TEST_USERS: {[key: string]: UserInfo} = {
   '00000000-0000-0000-0000-00000001': {
+    userId: '00000000-0000-0000-0000-00000001',
     firstName: 'Authority',
     lastName: 'Admin',
     roles: [
@@ -15,6 +16,7 @@ export let TEST_USERS: {[key: string]: UserInfo} = {
     organizationMdsId: 'MDSL1111AA',
   },
   '00000000-0000-0000-0000-00000002': {
+    userId: '00000000-0000-0000-0000-00000002',
     firstName: 'Authority',
     lastName: 'User',
     roles: ['AUTHORITY_USER', 'PARTICIPANT_USER'],
@@ -23,6 +25,7 @@ export let TEST_USERS: {[key: string]: UserInfo} = {
     organizationMdsId: 'MDSL1111AA',
   },
   '00000000-0000-0000-0000-00000003': {
+    userId: '00000000-0000-0000-0000-00000003',
     firstName: 'Participant',
     lastName: 'Admin',
     roles: ['PARTICIPANT_ADMIN', 'PARTICIPANT_CURATOR', 'PARTICIPANT_USER'],
@@ -31,6 +34,7 @@ export let TEST_USERS: {[key: string]: UserInfo} = {
     organizationMdsId: 'MDSL2222BB',
   },
   '00000000-0000-0000-0000-00000004': {
+    userId: '00000000-0000-0000-0000-00000004',
     firstName: 'Participant',
     lastName: 'User',
     roles: ['PARTICIPANT_USER'],
@@ -39,6 +43,7 @@ export let TEST_USERS: {[key: string]: UserInfo} = {
     organizationMdsId: 'MDSL2222BB',
   },
   '00000000-0000-0000-0000-00000005': {
+    userId: '00000000-0000-0000-0000-00000005',
     firstName: 'Created',
     lastName: 'User',
     roles: [],
@@ -47,6 +52,7 @@ export let TEST_USERS: {[key: string]: UserInfo} = {
     organizationMdsId: '',
   },
   '00000000-0000-0000-0000-00000006': {
+    userId: '00000000-0000-0000-0000-00000006',
     firstName: 'Pending',
     lastName: 'User',
     roles: [],