+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Initial Authority Portal MVP Release with a minimalistic feature set.
 - Add endpoints to deactivate users
 - Add endpoints to reactivate users
 - Unregister own connectors
+- Add userId to /api/user-info endpoint
 
 #### Patch